+++ audio/downmixer_test.go
@@ -8,12 +8,12 @@ import (
 )
 
 func TestDownmixer(t *testing.T) {
-	samples := []float64{0.1, 0.1, 0.2, 0.2, 0.3, 0.3}
-	want := []float64{0.1, 0.2, 0.3}
+	samples := []float32{0.1, 0.1, 0.2, 0.2, 0.3, 0.3}
+	want := []float32{0.1, 0.2, 0.3}
 
 	downmix := audio.NewDownmixer(audio.NewReader(samples), 2)
 
-	got := make([]float64, len(samples)/2)
+	got := make([]float32, len(samples)/2)
 	_, err := downmix.ReadSamples(got)
 	if err != nil {
 		t.Fatal(err)
