+++ audio/mixer_test.go
@@ -8,13 +8,13 @@ import (
 )
 
 func TestMixer(t *testing.T) {
-	src1 := audio.NewReader([]float64{0.1, 0.2, 0.3, 0.0})
-	src2 := audio.NewReader([]float64{0.2, 0.3, 0.4, 0.0})
-	want := []float64{0.3, 0.5, 0.7, 0.0}
+	src1 := audio.NewReader([]float32{0.1, 0.2, 0.3, 0.0})
+	src2 := audio.NewReader([]float32{0.2, 0.3, 0.4, 0.0})
+	want := []float32{0.3, 0.5, 0.7, 0.0}
 
 	mixer := audio.NewMixer(src1)
 	mixer.Add(src2)
-	got := make([]float64, len(want))
+	got := make([]float32, len(want))
 	_, err := mixer.ReadSamples(got)
 	if err != nil {
 		t.Fatal(err)
