+++ encoding/g711/alaw_test.go
@@ -8,7 +8,7 @@ import (
 )
 
 func TestAlawRoundTrip(t *testing.T) {
-	samples := []float64{0, 0.5, -0.5, 1, -1}
+	samples := []float32{0, 0.5, -0.5, 1, -1}
 
 	encoded := g711.EncodeAlaw(samples)
 	decoded := g711.DecodeAlaw(encoded)
